% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hilandyn_map.R
\name{hilandyn_map}
\alias{hilandyn_map}
\title{High-dimensional detection of Landscape Dynamics}
\usage{
hilandyn_map(
  sr_data,
  si_data,
  nob_data = NULL,
  out_path = NULL,
  sr_ind = NULL,
  si_ind = NULL,
  nob_ind = NULL,
  years,
  win_side = 3,
  cell_weights = TRUE,
  cng_dir,
  th_const = 1,
  noise_iter_max = 2,
  nob_init_min = 5,
  rmse = TRUE,
  use_last = TRUE,
  expand = TRUE,
  roi_vec = NULL,
  clip_input = FALSE,
  n_copy = 4,
  cores = 1
)
}
\arguments{
\item{sr_data}{character string. Surface reflectance data. Either the name of a \code{SpatRaster} object or the folder where rasters are stored. If \code{NULL} only spectral indices are used. See details for more information.}

\item{si_data}{character string. Spectral indices data. Either the name of a \code{SpatRaster} object or the folder where rasters are stored. If \code{NULL} only reflectance bands are used. See details for more information.}

\item{nob_data}{character string. Number of clear observations available per-pixel for producing reflectance composites. Either the name of a \code{SpatRaster} object or the folder where rasters are stored. If \code{NULL} a dummy value of 999 is used. See details for more information.}

\item{out_path}{character. The path where output rasters (in \emph{.tif} format) will be saved. Folders are created recursively if they do not exist. If \code{NULL} the output is a \code{SpatRaster} object created by the \pkg{terra} package.}

\item{sr_ind}{numeric vector. Indices of the layers containing reflectance data in single-date input rasters. If \code{NULL} all layers are used. Ignored if \code{sr_data} is a \code{SpatRaster}.}

\item{si_ind}{numeric vector. Indices of the layers containing spectral indices data in single-date input rasters. If \code{NULL} all layers are used. Ignored if \code{si_data} is a \code{SpatRaster}.}

\item{nob_ind}{numeric vector. Index of the layer containing data relative to the number of observations in single-date input rasters. If \code{NULL} the first layer of the input raster is used. Ignored if \code{nob_data} is a \code{SpatRaster}.}

\item{years}{numeric vector. Time interval (years) covered by the input rasters.}

\item{win_side}{integer. Width (in cells) of the spatial kernel used to extract input data from rasters. Must be an odd number.}

\item{cell_weights}{logical. Enable or disable the use of cell-based weights within the spatial kernel.}

\item{cng_dir}{numeric vector. Direction (either 1 or -1) of change in each spectral variable associated to the occurrence of a forest disturbance. It is computed as the difference between pre-disturbance and post-disturbance values.}

\item{th_const}{numeric. Constant value controlling the sensitivity to deviations from linearity in the HiTS procedure. Typical values are comprised in the interval \eqn{[0.7, 1.3]}.}

\item{noise_iter_max}{integer. The maximum number of iterations allowed for removing impulsive noise with the noise filter. The noise filter is disabled when the value is 0.}

\item{nob_init_min}{integer. The minimum number of clear observations available per time step in the first two time steps of the time series.}

\item{rmse}{logical. Determines whether to compute the root mean square error for each band in the focal cell. Original and estimated values are processed before noise filtering and gap filling.}

\item{use_last}{logical. Determines whether changepoints detected at the last time point are ignored or not.}

\item{expand}{logical. Whether to expand the raster by adding virtual rows and columns outside of it. New cells are filled with \code{NA}.}

\item{roi_vec}{character. Path to a spatial vector file containing polygons, e.g. a shapefile, of the region of interest. It is read as a \code{SpatVector} object using the \pkg{terra} package, and is used for masking input raster data.}

\item{clip_input}{logical. Whether to clip the input rasters using the extent of the region of interest.}

\item{n_copy}{integer. Number of copies of the input raster required for generating the output. Increasing this value will cause a reduction in the size of raster chucks that used during computation. Recommended values ranges between 1 and 4.}

\item{cores}{integer. Number of CPU cores employed for parallelising the analysis.}
}
\value{
If \code{out_path} is \code{NULL}, a \code{SpatRaster} containing the following layers.
\item{D_MAX_YR}{Year corresponding to the disturbance event with the maximum spectral change magnitude \code{D_MAX_MG} (single value).}
\item{D_MAX_MG}{Maximum spectral change magnitude of any disturbance computed as the median of bands and cells within the spatial kernel (single value).}
\item{D_FRS_YR}{Year corresponding to the first disturbance event (single value).}
\item{D_FRS_MG}{Spectral change magnitude of the first disturbance event computed as the median of bands and cells within the spatial kernel (single value).}
\item{D_LST_YR}{Year corresponding to the last disturbance event (single value).}
\item{D_LST_MG}{Spectral change magnitude of the last disturbance event computed as the median of bands and cells within the spatial kernel (single value).}
\item{D_NUM}{Total number of disturbance events (single value).}
\item{G_MAX_YR}{Year corresponding to the greening event with the maximum spectral change magnitude \code{G_MAX_MG} (single value).}
\item{G_MAX_MG}{Maximum spectral change magnitude of greening computed as the median of bands and cells within the spatial kernel (single value).}
\item{G_FRS_YR}{Year corresponding to the first greening event (single value).}
\item{G_FRS_MG}{Spectral change magnitude of the first greening event computed as the median of bands and cells within the spatial kernel (single value).}
\item{G_LST_YR}{Year corresponding to the last disturbance event (single value).}
\item{G_LST_MG}{Spectral change magnitude of the last disturbance event computed as the median of bands and cells within the spatial kernel (single value).}
\item{G_NUM}{Total number of greening events (single value).}
\item{N_GAP}{Number of gaps in the time series, if any (single value).}
\item{N_NOISE}{Number of years containing impulsive noise, if any (single value).}
\item{RMSE}{Root mean square error of each band in the focal cell (one value per band).}
\item{LEN}{Length of segments (one value per year).}
\item{CPT_ID}{Type of change (one value per year). One of the following values: 101 (disturbance); 102 (greening); 9 (other change). Otherwise 0 for no change.}
\item{CPT_FOC}{Number of changepoints detected in the focal cell. The minimum value is zero and the maximum corresponds to the number of bands.}
\item{NOISE}{Position of the impulsive noise in the time series (one value per year).}
\item{MED_REL_MAG}{Median relative magnitude computed using all the cells in the spatial kernel and the bands.}
\item{EST}{Estimated values of the bands in the focal cell (one value per year and band).}
\item{SLO}{Slope of the linear segments of the bands in the focal cell (one value per year and band).}
\item{MAG}{Magnitude in absolute terms of the bands in the focal cell (one value per year and band).}
\item{REL_MAG}{Magnitude in relative terms of the bands in the focal cell (one value per year and band).}
}
\description{
This is the main function of the package.
}
\details{
\code{hilandyn_map()} produces maps relative to landscape dynamics through the segmentation of high-dimensional Landsat time series.
These latter include information from the spatial and spectral domains and are analysed using the High-dimensional Trend Segmentation (HiTS) procedure proposed by \insertCite{maeng2019adaptive;textual}{hilandyn}.
The HiTS procedure aims to detect changepoints in a piecewise linear signal where their number and location are unknown. Changes can occur in the intercept, slope or both of linear trends.
High-dimensional time series can include single or multiple spectral bands/indices, hereafter referred to as bands.
Impulsive noise, i.e. outliers in the time series, are removed through an iterative procedure.
One-year gaps in the time series are filled using either linear interpolation or extrapolation.
If input data, i.e. \code{sr_data}, \code{si_data}, and \code{nob_data}, are provided as paths to folders, one multiband raster in \emph{.tif} format per year is required. Raster file names should include the year such that they can be ordered by time.
Input rasters are read as \code{SpatRaster} objects created by the \pkg{terra} package.
}
\examples{
library(hilandyn)

# Load raster data
data(lnd_sr)
data(lnd_si)
lnd_sr <- terra::rast(lnd_sr)
lnd_si <- terra::rast(lnd_si)

# Process data
rsout <- hilandyn_map(sr_data = lnd_sr,
                      si_data = lnd_si,
                      nob_data = NULL,
                      years = 1985:2020,
                      cng_dir = c(-1, -1, 1, 1),
                      cores = 1)

# Plot the maximum disturbance magnitude and the corresponding year
terra::plot(rsout[["D_MAX_MG"]])
terra::plot(rsout[["D_MAX_YR"]])


}
\references{
\insertRef{maeng2019adaptive}{hilandyn}
}
\seealso{
\code{\link{hilandyn_int}}
}
\author{
Donato Morresi, \email{donato.morresi@unito.it}
}
