% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnd_sr.R
\docType{data}
\name{lnd_sr}
\alias{lnd_sr}
\title{Landsat time series of surface reflectance.}
\format{
A \code{PackedSpatRaster} object with 36 layers, one per year.
}
\usage{
lnd_sr
}
\description{
A raster dataset containing annual surface reflectance in the SWIR2 band from 1985 to 2020.
The spatial subset covers a wildfire occurred in the Aosta Valley (Italy) during the winter of 2005.
Landsat imagery was acquired by the Thematic Mapper (TM), Enhanced Thematic Mapper + (ETM+) and Operational Land Imager (OLI)
sensors. Surface reflectance was derived from Level 1 Tier 1 imagery from Landsat Collection 1 in the USGS archive using the
Framework for Operational Radiometric Correction for Environmental monitoring (FORCE) software \insertCite{frantz2019force}{hilandyn}.
Annual reflectance composites were produced using all the images acquired during the growing season (between June 1 and September 30).
The geometric median approach \insertCite{roberts2017high}{hilandyn} was employed to generate reflectance composites at the pixel level.
}
\references{
\insertRef{frantz2019force}{hilandyn}

\insertRef{roberts2017high}{hilandyn}
}
\keyword{datasets}
