% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hilandyn_int.R
\name{hilandyn_int}
\alias{hilandyn_int}
\title{High-dimensional Detection of Landscape Dynamics engine}
\usage{
hilandyn_int(
  x,
  nob,
  nr,
  ny,
  nb,
  nc,
  yrs,
  foc_ind,
  cell_weights = TRUE,
  ev = NA,
  cng_dir,
  th_const = 1,
  noise_iter_max = 2,
  nob_init_min = 5,
  use_last = TRUE,
  as_list = FALSE
)
}
\arguments{
\item{x}{numeric matrix. Input time series where rows contain time-ordered data relative to individual bands for each cell in the spatial kernel.}

\item{ny}{integer. Number of time steps (years) in the time series.}

\item{nb}{integer. Number of input bands in the time series.}

\item{nc}{integer. The number of cells within the spatial kernel.}

\item{yrs}{numeric vector. The sequence of years to be analysed.}

\item{foc_ind}{numeric vector. Row indices of the bands in the focal cell.}

\item{cell_weights}{logical. Whether to compute cell-based weights within the spatial kernel.}

\item{ev}{logical vector. A vector containing \code{NA} values to be used in case of processing failure.}

\item{cng_dir}{numeric vector. Direction (either 1 or -1) of the spectral change caused by a disturbance in each band. It is computed as the difference between pre-disturbance and post-disturbance values.}

\item{th_const}{numeric. Constant value controlling the sensitivity to deviations from linearity in the HiTS procedure. Typical values are comprised in the interval \eqn{[0.7, 1.3]}.}

\item{noise_iter_max}{integer. The maximum number of iterations allowed for removing impulsive noise with the noise filter. The noise filter is disabled when the value is 0.}

\item{nob_init_min}{integer. The minimum number of clear observations available per time step in the first two time steps of the time series.}

\item{use_last}{logical. Determines whether changepoints detected at the last time point are ignored or not.}

\item{as_list}{logical. If \code{TRUE} the output is a \code{list}. Otherwise, the output is a \code{numeric} \code{vector}.}
}
\value{
If \code{as_list} is \code{TRUE}, a \code{list} containing the following elements. Otherwise the result is a \code{numeric} \code{vector}.
\item{est}{Estimated values (one value for each year and band).}
\item{cpt}{Detected changepoints (one value for each year and band).}
\item{slo}{Slope of the linear segments (one value for each year and band).}
\item{mag}{Magnitude in absolute terms (one value for each year and band).}
\item{mag_rel}{Magnitude in relative terms (one value for each year and band).}
\item{len}{Length of segments (one value per year).}
\item{cpt_id}{Type of change (one value per year). One of the following values: 101 (abrupt disturbance); 102 (abrupt greening); 9 (other change). Otherwise \code{NA} for no change.}
\item{cpt_foc}{Number of changepoints detected by the HiTS procedure in the focal cell. The minimum value is zero and the maximum corresponds to the number of bands.}
\item{noise}{Impulsive noise (one value per year).}
\item{d_dur}{Duration of disturbance (one value per year).}
\item{g_dur}{Duration of greening (one value per year).}
\item{d_max}{Maximum disturbance change magnitude (in relative tems) throughout the time series (one value per band).}
\item{d_fst}{Change magnitude (in relative tems) associated with the first disturbance detected within the time series (one value per band).}
\item{g_max}{Maximum greening change magnitude (in relative tems) throughout the time series (one value per band).}
\item{d_max_md}{Median among bands using values of \code{D_MAX} (single value).}
\item{d_fst_md}{Median among bands using values of \code{D_FST} (single value).}
\item{g_max_md}{Median among bands using values of \code{G_MAX} (single value).}
\item{d_max_yr}{Year corresponding to \code{D_MAX_MD} (single value).}
\item{d_fst_yr}{Year corresponding to \code{D_FST_MD} (single value).}
\item{g_max_yr}{Year corresponding to \code{G_MAX_MD} (single value).}
\item{d_max_dr}{Duration of the disturbance corresponding to \code{D_MAX_MD} (single value).}
\item{d_fst_dr}{Duration of the disturbance corresponding to \code{D_FST_MD} (single value).}
\item{g_max_dr}{Duration of the greening corresponding to \code{G_MAX_MD} (single value).}
\item{d_max-id}{Type of change (\code{CPT_ID}) of the disturbance corresponding to \code{D_MAX_MD} (single value).}
\item{d_fst_id}{Type of change (\code{CPT_ID}) of the disturbance corresponding to \code{D_FST_MD} (single value).}
\item{g_max_id}{Type of change (\code{CPT_ID}) of the greening corresponding to \code{G_MAX_MD} (single value).}
\item{n_gap}{Number of gaps in the time series, if any (single value).}
\item{n_noise}{Number of years containing impulsive noise, if any (single value).}
}
\description{
This is the internal function called by \code{\link{hilandyn_map}}.
}
\details{
Typically, \code{hilandyn_int()} is not directly called by the user.
It analyses inter-annual high-dimensional time series to detect changes in spectral trends within a spatial kernel.
High-dimensional time series can include single or multiple spectral bands/indices, hereafter referred to as bands.
Input data can be a \code{numeric} \code{vector} extracted from a \code{SpatRaster} or a \code{matrix}.
Impulsive noise, i.e. outliers in the time series, can be removed through an iterative procedure.
One-year gaps in the time series are filled using either linear interpolation or extrapolation.
Changes in the intercept, slope or both of linear trends are detected using the High-dimensional Trend Segmentation (HiTS) procedure proposed by \insertCite{maeng2019adaptive;textual}{hilandyn}.
The HiTS procedure aims at detecting changepoints in a piecewise linear signal where their number and location are unknown.
}
\examples{
library(hilandyn)

# Load raster data
data(lnd_si)
lnd_si <- terra::rast(lnd_si)

# Extract values
v <- terra::values(lnd_si)[2000,]

# Process data
out <- hilandyn_int(v,
                    nb = 3,
                    ny = 36,
                    nc  = 9,
                    yrs = 1985:2020,
                    foc_ind = c(5, 14, 23),
                    cell_weights = TRUE,
                    cng_dir = c(-1, 1, 1),
                    as_list = TRUE)
                
# Plot results
bands <- c("MSI", "TCW", "TCA")
years <- 1985:2020
par(mfrow=c(length(bands), 1))

for(i in seq_along(bands)) {
plot(matrix(v[foc_ind, ], nrow = length(bands), ncol = length(years))[i,], type = "l", col = 2,
       lwd = 2, ylab = "", xlab = "", xaxt = "n")
lines(out$est[i,], type = "l", col = 4, lwd = 2, lty = 1)
abline(v = which(out$cpt_id \%in\% c(101, 201)), lty = 2, col = 2, lwd = 2)
axis(1, at = seq_along(years), labels = years)
title(main = paste(bands[i]), adj = 0)
}

}
\references{
\insertRef{maeng2019adaptive}{hilandyn}
}
\author{
Donato Morresi, \email{donato.morresi@gmail.com}
}
