% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hilandyn_map.R
\name{hilandyn_map}
\alias{hilandyn_map}
\title{High-dimensional Detection of Landscape Dynamics}
\usage{
hilandyn_map(
  sr_data,
  si_data,
  nob_data,
  out_path = NULL,
  sr_ind = NULL,
  si_ind = NULL,
  nob_ind = NULL,
  years,
  win_side = 3,
  cell_weights = TRUE,
  cng_dir,
  th_const = 1,
  noise_iter_max = 2,
  nob_init_min = 5,
  use_last = TRUE,
  expand = TRUE,
  roi_vec = NULL,
  clip_input = FALSE,
  n_copy = 4,
  cores = 1
)
}
\arguments{
\item{sr_data}{character string. Surface reflectance data. Either the name of a \code{SpatRaster} object or the folder where rasters are stored. If \code{NULL} only spectral indices are used. See details for more information.}

\item{si_data}{character string. Spectral indices data. Either the name of a \code{SpatRaster} object or the folder where rasters are stored. If \code{NULL} only reflectance bands are used. See details for more information.}

\item{nob_data}{character string. Number of clear observations available per-pixel for producing reflectance composites. Either the name of a \code{SpatRaster} object or the folder where rasters are stored. If \code{NULL} a dummy value of 999 is used. See details for more information.}

\item{out_path}{character. The path where output rasters (in \emph{.tif} format) will be saved. Folders are created recursively if they do not exist. If \code{NULL} the output is a \code{SpatRaster} object created by the \pkg{terra} package.}

\item{sr_ind}{numeric vector. Indices of the layers containing reflectance data in single-date input rasters. If \code{NULL} all layers are used. Ignored if \code{sr_data} is a \code{SpatRaster}.}

\item{si_ind}{numeric vector. Indices of the layers containing spectral indices data in single-date input rasters. If \code{NULL} all layers are used. Ignored if \code{si_data} is a \code{SpatRaster}.}

\item{nob_ind}{numeric vector. Index of the layer containing data relative to the number of observations in single-date input rasters. If \code{NULL} the first layer of the input raster is used. Ignored if \code{nob_data} is a \code{SpatRaster}.}

\item{years}{numeric vector. Time interval (years) covered by the input rasters.}

\item{win_side}{integer. Width (in cells) of the spatial kernel used to extract input data from rasters. Must be an odd number.}

\item{cell_weights}{logical. Enable or disable the use of cell-based weights within the spatial kernel.}

\item{cng_dir}{numeric vector. Direction (either 1 or -1) of change in each spectral variable associated to the occurrence of a forest disturbance. It is computed as the difference between pre-disturbance and post-disturbance values.}

\item{th_const}{numeric. Constant value controlling the sensitivity to deviations from linearity in the HiTS procedure. Typical values are comprised in the interval \eqn{[0.7, 1.3]}.}

\item{noise_iter_max}{integer. The maximum number of iterations allowed for removing impulsive noise with the noise filter. The noise filter is disabled when the value is 0.}

\item{nob_init_min}{integer. The minimum number of clear observations available per time step in the first two time steps of the time series.}

\item{use_last}{logical. Determines whether changepoints detected at the last time point are ignored or not.}

\item{expand}{logical. Whether to expand the raster by adding virtual rows and columns outside of it. New cells are filled with \code{NA}.}

\item{roi_vec}{character. Path to a spatial vector file containing polygons, e.g. a shapefile, of the region of interest. It is read as a \code{SpatVector} object using the \pkg{terra} package, and is used for masking input raster data.}

\item{clip_input}{logical. Whether to clip the input rasters using the extent of the region of interest.}

\item{n_copy}{integer. Number of copies of the input raster required for generating the output. Increasing this value will cause a reduction in the size of raster chucks that used during computation. Recommended values ranges between 1 and 4.}

\item{cores}{integer. Number of CPU cores employed for parallelising the analysis.}
}
\value{
If \code{out_path} is \code{NULL}, a \code{SpatRaster} containing the following layers.
\item{EST}{Estimated values (one layer for each year and band).}
\item{CPT}{Detected changepoints (one layer for each year and band).}
\item{SLO}{Slope of the linear segments (one layer for each year and band).}
\item{MAG}{Magnitude in absolute terms (one layer for each year and band).}
\item{MAG_REL}{Magnitude in relative terms (one layer for each year and band).}
\item{LEN}{Length of segments (one layer per year).}
\item{CPT_ID}{Type of change (one layer per year). One of the following values: 101 (abrupt disturbance); 102 (abrupt greening); 201 (gradual disturbance); 202 (gradual greening); 9 (other change). Otherwise \code{NA} for no change.}
\item{CPT_FOC}{Number of changepoints detected by the HiTS procedure in the focal cell. The minimum value is zero and the maximum corresponds to the number of bands.}
\item{NOISE}{Impulsive noise (one layer per year).}
\item{D_DUR}{Duration of disturbance (one layer per year).}
\item{G_DUR}{Duration of greening (one layer per year).}
\item{D_MAX}{Maximum disturbance change magnitude (in relative terms) throughout the time series (one layer per band).}
\item{D_FST}{Change magnitude (in relative terms) associated with the first disturbance detected within the time series (one layer per band).}
\item{G_MAX}{Maximum greening change magnitude (in relative tems) throughout the time series (one layer per band).}
\item{D_MAX_MD}{Median among bands using values of \code{D_MAX} (single layer).}
\item{D_FST_MD}{Median among bands using values of \code{D_FST} (single layer).}
\item{G_MAX_MD}{Median among bands using values of \code{G_MAX} (single layer).}
\item{D_MAX_YR}{Year corresponding to \code{D_MAX_MD} (single layer).}
\item{D_FST_YR}{Year corresponding to \code{D_FST_MD} (single layer).}
\item{G_MAX_YR}{Year corresponding to \code{G_MAX_MD} (single layer).}
\item{D_MAX_DR}{Duration of the disturbance corresponding to \code{D_MAX_MD} (single layer).}
\item{D_FST_DR}{Duration of the disturbance corresponding to \code{D_FST_MD} (single layer).}
\item{G_MAX_DR}{Duration of the greening corresponding to \code{G_MAX_MD} (single layer).}
\item{D_MAX_ID}{Type of change (\code{CPT_ID}) of the disturbance corresponding to \code{D_MAX_MD} (single layer).}
\item{D_FST_ID}{Type of change (\code{CPT_ID}) of the disturbance corresponding to \code{D_FST_MD} (single layer).}
\item{G_MAX_ID}{Type of change (\code{CPT_ID}) of the greening corresponding to \code{G_MAX_MD} (single layer).}
\item{N_GAP}{Number of gaps in the time series, if any (single layer).}
\item{N_NOISE}{Number of years containing impulsive noise, if any (single layer).}
}
\description{
This is the main function of the package.
}
\details{
\code{hilandyn_map()} produces maps relative to landscape dynamics through the segmentation of high-dimensional Landsat time series.
These latter include information in the spatial and spectral dimensions and are analysed using the High-dimensional Trend Segmentation (HiTS) procedure proposed by \insertCite{maeng2019adaptive;textual}{hilandyn}.
The HiTS procedure aims to detect changepoints in a piecewise linear signal where their number and location are unknown. Changes can occur in the intercept, slope or both of linear trends.
Time series can include single or multiple spectral bands/indices, hereafter referred to as bands.
Impulsive noise, i.e. outliers in the time series, are removed through an iterative procedure.
One-year gaps in the time series are filled using either linear interpolation or extrapolation.
If input data are provided as folders, one multiband raster in \emph{.tif} format per year should be present within each folder. Moreover, filenames should contain the year such that rasters can be ordered by year.
Input rasters are read as \code{SpatRaster} objects created by the \pkg{terra} package.
}
\examples{
library(hilandyn)

# Load raster data
data(lnd_sr)
data(lnd_si)
lnd_sr <- terra::rast(lnd_sr)
lnd_si <- terra::rast(lnd_si)

# Process data
rsout <- hilandyn(sr_data = "lnd_sr",
                  si_data = "lnd_si",
                  years = 1985:2020,
                  cng_dir = c(-1, -1, 1, 1),
                  cores = parallel::detectCores() - 1)

# Plot the maximum disturbance magnitude
terra::plot(rsout[["D_MAX_MD"]])

}
\references{
\insertRef{maeng2019adaptive}{hilandyn}
}
\seealso{
\code{\link{hilandyn_int}}
}
\author{
Donato Morresi, \email{donato.morresi@gmail.com}
}
